// next-750-year-sdam/src/types/StrapiImage.ts
export interface StrapiImage {
    id: number;
    documentId: string;
    name: string;
    alternativeText: null;
    caption: null;
    width: number;
    height: number;
    formats: Formats;
    hash: string;
    ext: string;
    mime: string;
    size: number;
    url: string;
    previewUrl: null;
    provider: string;
    provider_metadata: null;
    createdAt: Date;
    updatedAt: Date;
    publishedAt: Date;
}

interface Formats {
    thumbnail: Large;
    medium: Large;
    small: Large;
    large: Large;
}

interface Large {
    name: string;
    hash: string;
    ext: string;
    mime: string;
    path: null;
    width: number;
    height: number;
    size: number;
    sizeInBytes: number;
    url: string;
}